Refactor linters class herarcy
